requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
python-dotenv==1.0.0
certifi==2023.11.17
